requirements.txt
Django==3.1.1
djangorestframework==3.11.1
pylint==2.5.3
pip install django-rest-swagger==2.2.0

